pos-cloud/streamlit-app/requirements.txt
streamlit>=1.32.0
psycopg2-binary==2.9.9
groq>=0.11.0
plotly>=5.19.0
pandas>=2.0.0
